+++ bee-ternary/CHANGELOG.md
@@ -23,7 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Renamed feature `serde1` to `serde`;
+- Renamed feature `serde` to `serde`;
 
 ## 0.5.2 - 2021-11-19